Multiple very minor bugfixes and changes that came in a meeting at 1.6.2010
- Posts after name in front page
- RSS / Voice emulator feeds in front page
- Text changed to optional in content forms
- Some string lenghts changed
- Added files dir to repo
- Language selection checker for cnotent forms
etc

+++ application/models/Files.php
@@ -68,10 +68,14 @@ class Default_Model_Files extends Zend_Db_Table_Abstract
     	$hash = hash_hmac_file('sha1', $uploadedFile['tmp_name'], $id_cnt.$uploadedFile['name']);
 		$dir = "files/".$id_usr."/";
     	if (! file_exists($dir)) {
-    		mkdir($dir, 0700, true);
+    		mkdir($dir, 0777, true);
     	} 
-    	move_uploaded_file($uploadedFile['tmp_name'], $dir.$hash);
     	
+    	if ( file_exists($dir.$hash)) {
+    		return false;
+    	}
+    	
+    	move_uploaded_file($uploadedFile['tmp_name'], $dir.$hash);
     	$file = $this->createRow();
     	$file->id_cnt_fil = $id_cnt;
     	$file->id_usr_fil = $id_usr;
